[REF] l10n_br_fiscal: merge mixin and mixin methods files
